Remove redundant "as usize" keyword

+++ src/display/lca2021_badge.rs
@@ -101,7 +101,7 @@ pub fn connect(
     builder.spawn(move || {
         let display0 = get_display(bus.acquire_i2c(), 0x3C).unwrap();
         let display1 = get_display(bus.acquire_i2c(), 0x3D).unwrap();
-        let mut displays: [_; NUM_DISPLAYS as usize] = [display0, display1];
+        let mut displays: [_; NUM_DISPLAYS] = [display0, display1];
         display_thread::<_, NUM_PAGES, NUM_DISPLAYS>(tx_main, &mut displays, rx);
     })?;
